Move feedback components into teacher-feedback folder.
