+++ src/components/teacher-feedback/sequence-activity-feedback-badge.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import TeacherFeedbackBadge from "../../assets/svg-icons/teacher-feedback-badge.svg";
 
-import "./feedback-badge.scss";
+import "./sequence-activity-feedback-badge.scss";
 
 export const FeedbackBadge = () => {
   return (
